+++ src/features/dashboard/actions/DeleteNote.vue
