+++ src/features/dashboard/views/DashboardPage.vue
@@ -244,9 +244,9 @@ import { formatDate } from "@/utils/date";
 import { useUserStore } from "@/store/user";
 import { storeToRefs } from "pinia";
 import { ButtonNormal } from "lorga-ui";
-import DashboardCreateNote from "@/features/dashboard/actions/DashboardCreateNote.vue";
-import DashboardDeleteNote from "@/features/dashboard/actions/DashboardDeleteNote.vue";
-import DashboardUpdateNote from "@/features/dashboard/actions/DashboardUpdateNote.vue";
+import DashboardCreateNote from "@/features/dashboard/actions/CreateNote.vue";
+import DashboardDeleteNote from "@/features/dashboard/actions/DeleteNote.vue";
+import DashboardUpdateNote from "@/features/dashboard/actions/UpdateNote.vue";
 import { useNotes } from "../api/useNotes";
 import { useFollowUps } from "../api/useFollowUps";
 import { useRecords } from "../api/useRecords";
